Extracted testing code from app.helper.ts
