+++ src/app/component/about/pages/ethics/ethics.component.spec.ts
@@ -3,9 +3,9 @@ import {
   HttpTestingController
 } from "@angular/common/http/testing";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-import { testAppInitializer } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { AppConfigService } from "src/app/services/app-config/app-config.service";
+import { testAppInitializer } from "src/app/test.helper";
 import { EthicsComponent } from "./ethics.component";
 
 describe("AboutEthicsComponent", () => {
