+++ src/app/component/profile/pages/profile/their-profile.component.spec.ts
@@ -1,8 +1,8 @@
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
-import { testBawServices } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
+import { testBawServices } from "src/app/test.helper";
 import { TheirProfileComponent } from "./their-profile.component";
 
 describe("TheirProfileComponent", () => {
