+++ src/app/component/profile/pages/their-edit/their-edit.component.spec.ts
@@ -2,8 +2,9 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { FormlyModule } from "@ngx-formly/core";
-import { formlyRoot, testBawServices } from "src/app/app.helper";
+import { formlyRoot } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
+import { testBawServices } from "src/app/test.helper";
 import { TheirEditComponent } from "./their-edit.component";
 
 describe("TheirProfileEditComponent", () => {
