+++ src/app/component/projects/pages/assign/assign.component.spec.ts
@@ -1,8 +1,8 @@
 import { HttpClientModule } from "@angular/common/http";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
-import { testBawServices } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
+import { testBawServices } from "src/app/test.helper";
 import { AssignComponent } from "./assign.component";
 
 describe("AssignComponent", () => {
