+++ src/app/component/projects/pages/new/new.component.spec.ts
@@ -6,12 +6,13 @@ import {
 } from "@angular/core/testing";
 import { FormlyModule } from "@ngx-formly/core";
 import { Subject } from "rxjs";
-import { formlyRoot, testBawServices } from "src/app/app.helper";
+import { formlyRoot } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { mockProject, Project } from "src/app/models/Project";
 import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import { ProjectsService } from "src/app/services/baw-api/projects.service";
 import { NewComponent } from "./new.component";
+import { testBawServices } from "src/app/test.helper";
 
 describe("ProjectsNewComponent", () => {
   let api: ProjectsService;
