+++ src/app/component/security/pages/login/login.component.spec.ts
@@ -9,7 +9,7 @@ import { ActivatedRoute, Params, Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { FormlyModule } from "@ngx-formly/core";
 import { BehaviorSubject, Subject } from "rxjs";
-import { formlyRoot, testBawServices } from "src/app/app.helper";
+import { formlyRoot } from "src/app/app.helper";
 import { HomeComponent } from "src/app/component/home/home.component";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { SessionUser } from "src/app/models/User";
@@ -19,6 +19,7 @@ import {
   LoginDetails,
   SecurityService
 } from "src/app/services/baw-api/security.service";
+import { testBawServices } from "src/app/test.helper";
 import { LoginComponent } from "./login.component";
 
 describe("LoginComponent", () => {
