+++ src/app/component/security/pages/register/register.component.spec.ts
@@ -8,9 +8,10 @@ import {
 import { Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { FormlyModule } from "@ngx-formly/core";
-import { formlyRoot, testBawServices } from "src/app/app.helper";
+import { formlyRoot } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { SecurityService } from "src/app/services/baw-api/security.service";
+import { testBawServices } from "src/app/test.helper";
 import { RegisterComponent } from "./register.component";
 
 describe("RegisterComponent", () => {
