+++ src/app/component/shared/cms/cms.component.spec.ts
@@ -3,10 +3,10 @@ import {
   HttpTestingController
 } from "@angular/common/http/testing";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-import { testBawServices } from "src/app/app.helper";
 import { SessionUser } from "src/app/models/User";
 import { AppConfigService } from "src/app/services/app-config/app-config.service";
 import { SecurityService } from "src/app/services/baw-api/security.service";
+import { testBawServices } from "src/app/test.helper";
 import { SharedModule } from "../shared.module";
 import { CmsComponent } from "./cms.component";
