+++ src/app/component/shared/header/header.component.spec.ts
@@ -12,11 +12,11 @@ import { Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { BehaviorSubject, Subject } from "rxjs";
-import { testBawServices } from "src/app/app.helper";
 import { User } from "src/app/models/User";
 import { AppConfigService } from "src/app/services/app-config/app-config.service";
 import { SecurityService } from "src/app/services/baw-api/security.service";
 import { UserService } from "src/app/services/baw-api/user.service";
+import { testBawServices } from "src/app/test.helper";
 import { contactUsMenuItem } from "../../about/about.menus";
 import { homeMenuItem } from "../../home/home.menus";
 import { myAccountMenuItem } from "../../profile/profile.menus";
