+++ src/app/component/shared/menu/menu.component.spec.ts
@@ -1,11 +1,10 @@
 import { HttpClientModule } from "@angular/common/http";
 import { DebugElement } from "@angular/core";
-import { ComponentFixture, fakeAsync, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ActivatedRoute } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { List } from "immutable";
 import { BehaviorSubject } from "rxjs";
-import { testBawServices } from "src/app/app.helper";
 import { isGuestPredicate, isLoggedInPredicate } from "src/app/app.menus";
 import {
   AnyMenuItem,
@@ -16,6 +15,7 @@ import {
 import { StrongRoute } from "src/app/interfaces/strongRoute";
 import { SessionUser } from "src/app/models/User";
 import { SecurityService } from "src/app/services/baw-api/security.service";
+import { testBawServices } from "src/app/test.helper";
 import { getText } from "src/testHelpers";
 import { SharedModule } from "../shared.module";
 import { MenuButtonComponent } from "./button/button.component";
