+++ src/app/component/shared/user-badges/user-badges.component.spec.ts
@@ -9,11 +9,11 @@ import { RouterTestingModule } from "@angular/router/testing";
 import { List } from "immutable";
 import { DateTime } from "luxon";
 import { Subject } from "rxjs";
-import { testBawServices } from "src/app/app.helper";
 import { Project } from "src/app/models/Project";
 import { Site } from "src/app/models/Site";
 import { User } from "src/app/models/User";
 import { AccountService } from "src/app/services/baw-api/account.service";
+import { testBawServices } from "src/app/test.helper";
 import { MenuModule } from "../menu/menu.module";
 import { UserBadgeComponent } from "./user-badge/user-badge.component";
 import { UserBadgesComponent } from "./user-badges.component";
