+++ src/app/component/sites/pages/new/new.component.spec.ts
@@ -8,13 +8,14 @@ import { ActivatedRoute } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { FormlyModule } from "@ngx-formly/core";
 import { BehaviorSubject, Subject } from "rxjs";
-import { formlyRoot, testBawServices } from "src/app/app.helper";
+import { formlyRoot } from "src/app/app.helper";
 import { SharedModule } from "src/app/component/shared/shared.module";
 import { Project } from "src/app/models/Project";
 import { mockSite, Site } from "src/app/models/Site";
 import { ApiErrorDetails } from "src/app/services/baw-api/api.interceptor.service";
 import { ProjectsService } from "src/app/services/baw-api/projects.service";
 import { SitesService } from "src/app/services/baw-api/sites.service";
+import { testBawServices } from "src/app/test.helper";
 import { NewComponent } from "./new.component";
 
 describe("SitesNewComponent", () => {
