+++ src/app/services/baw-api/account.service.spec.ts
@@ -5,8 +5,8 @@ import {
 import { fakeAsync, TestBed, tick } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { Subject } from "rxjs";
-import { testAppInitializer } from "src/app/app.helper";
 import { User } from "src/app/models/User";
+import { testAppInitializer } from "src/app/test.helper";
 import { AccountService } from "./account.service";
 import { ApiErrorDetails } from "./api.interceptor.service";
 import { BawApiService, Filters } from "./baw-api.service";
