+++ src/app/services/baw-api/api.interceptor.service.spec.ts
@@ -9,8 +9,8 @@ import {
   HttpTestingController
 } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
-import { testBawServices } from "src/app/app.helper";
 import { SessionUser } from "src/app/models/User";
+import { testBawServices } from "src/app/test.helper";
 import { AppConfigService } from "../app-config/app-config.service";
 import { ApiErrorDetails, BawApiInterceptor } from "./api.interceptor.service";
 import { SecurityService } from "./security.service";
