+++ src/app/services/baw-api/baw-api.service.spec.ts
@@ -6,9 +6,9 @@ import {
 } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { BehaviorSubject, Subject } from "rxjs";
-import { testAppInitializer } from "src/app/app.helper";
 import { AbstractModel } from "src/app/models/AbstractModel";
 import { SessionUser } from "src/app/models/User";
+import { testAppInitializer } from "src/app/test.helper";
 import { AppConfigService } from "../app-config/app-config.service";
 import { ApiErrorDetails, BawApiInterceptor } from "./api.interceptor.service";
 import {
